# soa_week4/openapi-comparison/1_APIBlueprint/generated-flask-server/openapi_server/models/book.py
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from openapi_server.models.base_model import Model
from openapi_server import util


class Book(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, id=None, title=None, author=None, isbn=None, genre=None, published_year=None, price=None):  # noqa: E501
        """Book - a model defined in OpenAPI

        :param id: The id of this Book.  # noqa: E501
        :type id: float
        :param title: The title of this Book.  # noqa: E501
        :type title: str
        :param author: The author of this Book.  # noqa: E501
        :type author: str
        :param isbn: The isbn of this Book.  # noqa: E501
        :type isbn: str
        :param genre: The genre of this Book.  # noqa: E501
        :type genre: str
        :param published_year: The published_year of this Book.  # noqa: E501
        :type published_year: float
        :param price: The price of this Book.  # noqa: E501
        :type price: float
        """
        self.openapi_types = {
            'id': float,
            'title': str,
            'author': str,
            'isbn': str,
            'genre': str,
            'published_year': float,
            'price': float
        }

        self.attribute_map = {
            'id': 'id',
            'title': 'title',
            'author': 'author',
            'isbn': 'isbn',
            'genre': 'genre',
            'published_year': 'publishedYear',
            'price': 'price'
        }

        self._id = id
        self._title = title
        self._author = author
        self._isbn = isbn
        self._genre = genre
        self._published_year = published_year
        self._price = price

    @classmethod
    def from_dict(cls, dikt) -> 'Book':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Book of this Book.  # noqa: E501
        :rtype: Book
        """
        return util.deserialize_model(dikt, cls)

    @property
    def id(self) -> float:
        """Gets the id of this Book.

        ID tự tăng, do server gán  # noqa: E501

        :return: The id of this Book.
        :rtype: float
        """
        return self._id

    @id.setter
    def id(self, id: float):
        """Sets the id of this Book.

        ID tự tăng, do server gán  # noqa: E501

        :param id: The id of this Book.
        :type id: float
        """
        if id is None:
            raise ValueError("Invalid value for `id`, must not be `None`")  # noqa: E501

        self._id = id

    @property
    def title(self) -> str:
        """Gets the title of this Book.


        :return: The title of this Book.
        :rtype: str
        """
        return self._title

    @title.setter
    def title(self, title: str):
        """Sets the title of this Book.


        :param title: The title of this Book.
        :type title: str
        """
        if title is None:
            raise ValueError("Invalid value for `title`, must not be `None`")  # noqa: E501

        self._title = title

    @property
    def author(self) -> str:
        """Gets the author of this Book.


        :return: The author of this Book.
        :rtype: str
        """
        return self._author

    @author.setter
    def author(self, author: str):
        """Sets the author of this Book.


        :param author: The author of this Book.
        :type author: str
        """
        if author is None:
            raise ValueError("Invalid value for `author`, must not be `None`")  # noqa: E501

        self._author = author

    @property
    def isbn(self) -> str:
        """Gets the isbn of this Book.

        Có thể null  # noqa: E501

        :return: The isbn of this Book.
        :rtype: str
        """
        return self._isbn

    @isbn.setter
    def isbn(self, isbn: str):
        """Sets the isbn of this Book.

        Có thể null  # noqa: E501

        :param isbn: The isbn of this Book.
        :type isbn: str
        """

        self._isbn = isbn

    @property
    def genre(self) -> str:
        """Gets the genre of this Book.


        :return: The genre of this Book.
        :rtype: str
        """
        return self._genre

    @genre.setter
    def genre(self, genre: str):
        """Sets the genre of this Book.


        :param genre: The genre of this Book.
        :type genre: str
        """
        allowed_values = ["fiction", "non", "sci", "history", "biography"]  # noqa: E501
        if genre not in allowed_values:
            raise ValueError(
                "Invalid value for `genre` ({0}), must be one of {1}"
                .format(genre, allowed_values)
            )

        self._genre = genre

    @property
    def published_year(self) -> float:
        """Gets the published_year of this Book.

        Có thể null  # noqa: E501

        :return: The published_year of this Book.
        :rtype: float
        """
        return self._published_year

    @published_year.setter
    def published_year(self, published_year: float):
        """Sets the published_year of this Book.

        Có thể null  # noqa: E501

        :param published_year: The published_year of this Book.
        :type published_year: float
        """

        self._published_year = published_year

    @property
    def price(self) -> float:
        """Gets the price of this Book.

        Có thể null  # noqa: E501

        :return: The price of this Book.
        :rtype: float
        """
        return self._price

    @price.setter
    def price(self, price: float):
        """Sets the price of this Book.

        Có thể null  # noqa: E501

        :param price: The price of this Book.
        :type price: float
        """

        self._price = price