# soa_week4/openapi-comparison/1_APIBlueprint/generated-flask-server/openapi_server/models/create_book_request.py
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from openapi_server.models.base_model import Model
from openapi_server import util


class CreateBookRequest(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, title=None, author=None, isbn=None, genre=None, published_year=None, price=None):  # noqa: E501
        """CreateBookRequest - a model defined in OpenAPI

        :param title: The title of this CreateBookRequest.  # noqa: E501
        :type title: str
        :param author: The author of this CreateBookRequest.  # noqa: E501
        :type author: str
        :param isbn: The isbn of this CreateBookRequest.  # noqa: E501
        :type isbn: str
        :param genre: The genre of this CreateBookRequest.  # noqa: E501
        :type genre: str
        :param published_year: The published_year of this CreateBookRequest.  # noqa: E501
        :type published_year: float
        :param price: The price of this CreateBookRequest.  # noqa: E501
        :type price: float
        """
        self.openapi_types = {
            'title': str,
            'author': str,
            'isbn': str,
            'genre': str,
            'published_year': float,
            'price': float
        }

        self.attribute_map = {
            'title': 'title',
            'author': 'author',
            'isbn': 'isbn',
            'genre': 'genre',
            'published_year': 'publishedYear',
            'price': 'price'
        }

        self._title = title
        self._author = author
        self._isbn = isbn
        self._genre = genre
        self._published_year = published_year
        self._price = price

    @classmethod
    def from_dict(cls, dikt) -> 'CreateBookRequest':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The CreateBookRequest of this CreateBookRequest.  # noqa: E501
        :rtype: CreateBookRequest
        """
        return util.deserialize_model(dikt, cls)

    @property
    def title(self) -> str:
        """Gets the title of this CreateBookRequest.


        :return: The title of this CreateBookRequest.
        :rtype: str
        """
        return self._title

    @title.setter
    def title(self, title: str):
        """Sets the title of this CreateBookRequest.


        :param title: The title of this CreateBookRequest.
        :type title: str
        """
        if title is None:
            raise ValueError("Invalid value for `title`, must not be `None`")  # noqa: E501

        self._title = title

    @property
    def author(self) -> str:
        """Gets the author of this CreateBookRequest.


        :return: The author of this CreateBookRequest.
        :rtype: str
        """
        return self._author

    @author.setter
    def author(self, author: str):
        """Sets the author of this CreateBookRequest.


        :param author: The author of this CreateBookRequest.
        :type author: str
        """
        if author is None:
            raise ValueError("Invalid value for `author`, must not be `None`")  # noqa: E501

        self._author = author

    @property
    def isbn(self) -> str:
        """Gets the isbn of this CreateBookRequest.


        :return: The isbn of this CreateBookRequest.
        :rtype: str
        """
        return self._isbn

    @isbn.setter
    def isbn(self, isbn: str):
        """Sets the isbn of this CreateBookRequest.


        :param isbn: The isbn of this CreateBookRequest.
        :type isbn: str
        """

        self._isbn = isbn

    @property
    def genre(self) -> str:
        """Gets the genre of this CreateBookRequest.

        Phải thuộc VALID_GENRES  # noqa: E501

        :return: The genre of this CreateBookRequest.
        :rtype: str
        """
        return self._genre

    @genre.setter
    def genre(self, genre: str):
        """Sets the genre of this CreateBookRequest.

        Phải thuộc VALID_GENRES  # noqa: E501

        :param genre: The genre of this CreateBookRequest.
        :type genre: str
        """
        if genre is None:
            raise ValueError("Invalid value for `genre`, must not be `None`")  # noqa: E501

        self._genre = genre

    @property
    def published_year(self) -> float:
        """Gets the published_year of this CreateBookRequest.


        :return: The published_year of this CreateBookRequest.
        :rtype: float
        """
        return self._published_year

    @published_year.setter
    def published_year(self, published_year: float):
        """Sets the published_year of this CreateBookRequest.


        :param published_year: The published_year of this CreateBookRequest.
        :type published_year: float
        """

        self._published_year = published_year

    @property
    def price(self) -> float:
        """Gets the price of this CreateBookRequest.


        :return: The price of this CreateBookRequest.
        :rtype: float
        """
        return self._price

    @price.setter
    def price(self, price: float):
        """Sets the price of this CreateBookRequest.


        :param price: The price of this CreateBookRequest.
        :type price: float
        """

        self._price = price