# soa_week4/openapi-comparison/1_APIBlueprint/generated-flask-server/openapi_server/models/pagination.py
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from openapi_server.models.base_model import Model
from openapi_server import util


class Pagination(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, page=None, limit=None, total=None, total_pages=None):  # noqa: E501
        """Pagination - a model defined in OpenAPI

        :param page: The page of this Pagination.  # noqa: E501
        :type page: float
        :param limit: The limit of this Pagination.  # noqa: E501
        :type limit: float
        :param total: The total of this Pagination.  # noqa: E501
        :type total: float
        :param total_pages: The total_pages of this Pagination.  # noqa: E501
        :type total_pages: float
        """
        self.openapi_types = {
            'page': float,
            'limit': float,
            'total': float,
            'total_pages': float
        }

        self.attribute_map = {
            'page': 'page',
            'limit': 'limit',
            'total': 'total',
            'total_pages': 'totalPages'
        }

        self._page = page
        self._limit = limit
        self._total = total
        self._total_pages = total_pages

    @classmethod
    def from_dict(cls, dikt) -> 'Pagination':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Pagination of this Pagination.  # noqa: E501
        :rtype: Pagination
        """
        return util.deserialize_model(dikt, cls)

    @property
    def page(self) -> float:
        """Gets the page of this Pagination.

        Trang hiện tại  # noqa: E501

        :return: The page of this Pagination.
        :rtype: float
        """
        return self._page

    @page.setter
    def page(self, page: float):
        """Sets the page of this Pagination.

        Trang hiện tại  # noqa: E501

        :param page: The page of this Pagination.
        :type page: float
        """

        self._page = page

    @property
    def limit(self) -> float:
        """Gets the limit of this Pagination.

        Số bản ghi mỗi trang (clamp 1–100)  # noqa: E501

        :return: The limit of this Pagination.
        :rtype: float
        """
        return self._limit

    @limit.setter
    def limit(self, limit: float):
        """Sets the limit of this Pagination.

        Số bản ghi mỗi trang (clamp 1–100)  # noqa: E501

        :param limit: The limit of this Pagination.
        :type limit: float
        """

        self._limit = limit

    @property
    def total(self) -> float:
        """Gets the total of this Pagination.

        Tổng số sách sau khi filter  # noqa: E501

        :return: The total of this Pagination.
        :rtype: float
        """
        return self._total

    @total.setter
    def total(self, total: float):
        """Sets the total of this Pagination.

        Tổng số sách sau khi filter  # noqa: E501

        :param total: The total of this Pagination.
        :type total: float
        """

        self._total = total

    @property
    def total_pages(self) -> float:
        """Gets the total_pages of this Pagination.

        max(1, ceil(total / limit))  # noqa: E501

        :return: The total_pages of this Pagination.
        :rtype: float
        """
        return self._total_pages

    @total_pages.setter
    def total_pages(self, total_pages: float):
        """Sets the total_pages of this Pagination.

        max(1, ceil(total / limit))  # noqa: E501

        :param total_pages: The total_pages of this Pagination.
        :type total_pages: float
        """

        self._total_pages = total_pages