# soa_week4/openapi-comparison/3_TypeSpec/generated-flask-server/openapi_server/controllers/books_controller.py
import connexion
from typing import Dict
from typing import Tuple
from typing import Union

from openapi_server.models.api_error import ApiError  # noqa: E501
from openapi_server.models.book import Book  # noqa: E501
from openapi_server.models.book_list_response import BookListResponse  # noqa: E501
from openapi_server.models.create_book_request import CreateBookRequest  # noqa: E501
from openapi_server.models.genre import Genre  # noqa: E501
from openapi_server.models.update_book_request import UpdateBookRequest  # noqa: E501
from openapi_server import util


def books_create(body):  # noqa: E501
    """books_create

    Tạo sách mới.  &#x60;id&#x60; &#x3D; &#x60;len(BOOKS) + 1&#x60; — không an toàn với concurrent requests, nhưng đây là in-memory demo.  Thiếu &#x60;title&#x60;, &#x60;author&#x60;, hoặc &#x60;genre&#x60; trong body sẽ raise Python KeyError → HTTP 500 (không có validation ở app layer). # noqa: E501

    :param create_book_request: 
    :type create_book_request: dict | bytes

    :rtype: Union[Book, Tuple[Book, int], Tuple[Book, int, Dict[str, str]]
    """
    create_book_request = body
    if connexion.request.is_json:
        create_book_request = CreateBookRequest.from_dict(connexion.request.get_json())  # noqa: E501
    return 'do some magic!'


def books_list(page=None, limit=None, genre=None, q=None):  # noqa: E501
    """books_list

    Lấy danh sách sách.  Thứ tự xử lý filter trong app.py: 1. Validate page, limit (parse int, clamp) 2. Validate genre (nếu có) 3. Filter theo genre (so sánh &#x3D;&#x3D;) 4. Filter theo q (substring, case-insensitive, trên title VÀ author) 5. Paginate # noqa: E501

    :param page: Số trang. Clamp bằng max(1, page) — giá trị &lt; 1 tự động thành 1. Non-integer → 400 INVALID_PARAM.
    :type page: int
    :param limit: Số bản ghi/trang. Clamp vào [1, 100]. Non-integer → 400 INVALID_PARAM.
    :type limit: int
    :param genre: Lọc chính xác theo thể loại (b[\&quot;genre\&quot;] &#x3D;&#x3D; genre). Nếu giá trị không thuộc Genre enum → 400 INVALID_PARAM.
    :type genre: dict | bytes
    :param q: Tìm kiếm trong title hoặc author. Strip whitespace, lowercase, substring match. Bỏ trống hoặc bỏ qua &#x3D; không filter.
    :type q: str

    :rtype: Union[BookListResponse, Tuple[BookListResponse, int], Tuple[BookListResponse, int, Dict[str, str]]
    """
    if connexion.request.is_json:
        genre =  Genre.from_dict(connexion.request.get_json())  # noqa: E501
    return 'do some magic!'


def by_id_delete(book_id):  # noqa: E501
    """by_id_delete

    Xóa sách khỏi in-memory store. Response 204 không có body (trả về empty string \&quot;\&quot;). # noqa: E501

    :param book_id: 
    :type book_id: str

    :rtype: Union[None, Tuple[None, int], Tuple[None, int, Dict[str, str]]
    """
    return 'do some magic!'


def by_id_get(book_id):  # noqa: E501
    """by_id_get

    Lấy thông tin chi tiết một cuốn sách.  book_id trong URL là string, app.py dùng BOOKS.get(book_id) — cần truyền đúng kiểu integer ID khi lookup. # noqa: E501

    :param book_id: 
    :type book_id: str

    :rtype: Union[Book, Tuple[Book, int], Tuple[Book, int, Dict[str, str]]
    """
    return 'do some magic!'


def by_id_update(book_id, body):  # noqa: E501
    """by_id_update

    Thay thế toàn bộ dữ liệu sách (PUT semantics).  Các field optional trong UpdateBookRequest sẽ fallback về giá trị cũ của sách nếu không có trong body: isbn       → data.get(\&quot;isbn\&quot;, book.get(\&quot;isbn\&quot;)) publishedYear → data.get(\&quot;publishedYear\&quot;, book.get(\&quot;publishedYear\&quot;)) price      → data.get(\&quot;price\&quot;, book.get(\&quot;price\&quot;)) # noqa: E501

    :param book_id: 
    :type book_id: str
    :param update_book_request: 
    :type update_book_request: dict | bytes

    :rtype: Union[Book, Tuple[Book, int], Tuple[Book, int, Dict[str, str]]
    """
    update_book_request = body
    if connexion.request.is_json:
        update_book_request = UpdateBookRequest.from_dict(connexion.request.get_json())  # noqa: E501
    return 'do some magic!'